requirements.txt
aiohttp==1.0.5
appdirs==1.4.3
async-timeout==1.1.0
chardet==2.3.0
discord.py==0.15.0
multidict==2.1.2
packaging==16.8
pyparsing==2.2.0
requests==2.12.1
six==1.10.0
websockets==3.2
